move the results client into testmonitor client
