--- docs/api_reference/result.rst
@@ -1,23 +1,0 @@
-.. _api_tag_page:
-
-nisystemlink.clients.result
-======================
-
-.. autoclass:: nisystemlink.clients.result.ResultClient
-   :exclude-members: __init__
-
-   .. automethod:: __init__
-   .. automethod:: create_results
-   .. automethod:: get_results
-   .. automethod:: query_results
-   .. automethod:: query_result_values
-   .. automethod:: update_results
-   .. automethod:: delete_result
-   .. automethod:: delete_results
-
-.. automodule:: nisystemlink.clients.result.models
-   :members:
-   :imported-members:
-
-.. automodule:: nisystemlink.clients.result.utilities
-   :members:
