+++ docs/api_reference/testmonitor.rst
@@ -7,6 +7,13 @@ nisystemlink.clients.testmonitor
    :exclude-members: __init__
 
    .. automethod:: __init__
+   .. automethod:: create_results
+   .. automethod:: get_results
+   .. automethod:: query_results
+   .. automethod:: query_result_values
+   .. automethod:: update_results
+   .. automethod:: delete_result
+   .. automethod:: delete_results
 
 .. automodule:: nisystemlink.clients.testmonitor.models
    :members:
