--- nisystemlink/clients/result/__init__.py
@@ -1,3 +1,0 @@
-from ._result_client import ResultClient
-
-# flake8: noqa
