+++ nisystemlink/clients/testmonitor/models/__init__.py
@@ -1,3 +1,13 @@
 from ._api_info import Operation, V2Operations, ApiInfo
+from ._result import Result, StatusObject, StatusType
+from ._results_partial_success import ResultsPartialSuccess
+from ._delete_results_partial_success import DeleteResultsPartialSuccess
+from ._paged_results import PagedResults
+from ._update_result_request import UpdateResultRequest
+from ._query_results_request import (
+    QueryResultsRequest,
+    ResultField,
+    QueryResultValuesRequest,
+)
 
 # flake8: noqa
